[fix] Fix O3 ReExec when used with Ruby.
Ruby would notify the O3 cpu that a load needs to be re-executed (checkSnoop()).
However, when the cache is blocked, the CPU still will push the load instruction
into blockedMemInst and re-execute it later when the cache is unblocked. This
causes an assertion failure in read().

I specifically do not try to call read() when the load instruction has ReExec
fault and has already been set Executed flag.

Change-Id: I3ad6393f4ef12775b895ade41cf9ba068db7e822