[fix] Fix the mnemonic for SSP macro instructions.
X86 decoder generator will split the macro-inst by '_' and use the first sub-string
as the mnemonic for the macro instruction. This makes all SSP macro-inst has a
mnemonic 'ssp'. We fix this by a special check in the macroop.isa so that the
ISAHandler can recognize the macroop.

Change-Id: Ie01dc1a3a01009d8a4a382b53d00da6b7a0e98a1
TODO: Recognize instruction by name sounds too hacky.