[fix] Do not mark load with variable latency.
Previously when implementing the LoadBlocked statistics, I mark the load with variable latency.
However, this causes the user instruction to be stalled until the load is committed and hurts
the performance. Therefore, I switched back to the original implementation.

This will make the LoadBlocked statistics unreal as a user instruction will never be blocked
by the load at issue stage. (It will be stalled at commit stage).

Change-Id: Idf39a56baf9737f4b78966af5cf5bb6c0e74e48a