[fix] Making the MessageBuffer between L1 and L2 not FIFO.
Previously I tried to fix the delayHead() of MessageBuffer, as when there are other
messages in the buffer that have later arrival time, delayHead() would violate the
FIFO ordering. I Fixed it by delaying all messages in the MessageBuffer. However,
this actually causes deadlock when the core is expecting some messages. Therefore,
I revert this.

Change-Id: Ibc20241eaf1af90471890c36409a65e6e71c7cd4