[misc] Record a known race condition for the coherence protocol.
There is a known race case for the coherence protocol when it receives:
L1_GETV at MT states. The problem is that we transit immediately to MT_IB,
and do not require the receiver to send Unblock. This has a problem if the
exclusive owner is also replacing that cache line.

Currently I just record it and has not try to fix it.

Change-Id: I5af38bdad8b9ff04057f92de5b10fcfc732f9968