[fix] Fix the bug of hitted stream request when in transient state.
Bug: When the cache line is in transient state, and we receive a offloaded stream request,
we consider it hit in MLC and immediately notify the MLC_SE. However, that is actually blocked
in the message buffer and will be processed later when the cache line transits back to stable
state. This causes the MLC_SE sees multiple requests to that stream slice.

Change-Id: Ic9adeda514f55b142696cc47243b46d4a3e21e01
Fix: Only notify the MLC_SE when the request gets served.