[feat] Support set latency for Ruby cache.
Although weird, we can now set the latency of each ruby cache level.
The implemenation of MESI_Three_Level_Cache only has latency for request and response, which are
tricky to map the lookup latency. So I charge the lookup latency at the request latency at upper
level cache. Hopefully this is OK.

Change-Id: I8d87ce0cca98f65899819930f0dae23a08a98e60