arch-x86: Fix default case in locked_opcodes.
The current implementation in locked_opcodes.isa uses
MultiInst to further switch on OPCODE_OP_BOTTOM3.
However, this will reach to M5_UNREACHABLE when there
is no match, and eventually a segment fault.
In my simulation, it is possible that the decoder is
decoding an invalid PC due to some control flow and
trigger this M5_UNREACHABLE.
This commit fixes that by returning an UD2 instruction
as other cases. This UD2 instruction will not be
discarded in most cases.

Change-Id: I24a766c8133bc56d9a3db753ee5891808a7cba35